Move some more bits
