+++ SpiNNaker-allocserv/src/main/java/uk/ac/manchester/spinnaker/alloc/bmp/FirmwareLoader.java
@@ -28,11 +28,11 @@ import static org.apache.commons.io.IOUtils.buffer;
 import static org.slf4j.LoggerFactory.getLogger;
 import static uk.ac.manchester.spinnaker.alloc.bmp.DataSectorTypes.BITFILE;
 import static uk.ac.manchester.spinnaker.alloc.bmp.DataSectorTypes.REGISTER;
+import static uk.ac.manchester.spinnaker.machine.board.FPGA.FPGA_ALL;
+import static uk.ac.manchester.spinnaker.machine.board.FPGA.FPGA_E_S;
+import static uk.ac.manchester.spinnaker.machine.board.FPGA.FPGA_N_NE;
+import static uk.ac.manchester.spinnaker.machine.board.FPGA.FPGA_SW_W;
 import static uk.ac.manchester.spinnaker.messages.Constants.WORD_SIZE;
-import static uk.ac.manchester.spinnaker.messages.model.FPGA.FPGA_ALL;
-import static uk.ac.manchester.spinnaker.messages.model.FPGA.FPGA_E_S;
-import static uk.ac.manchester.spinnaker.messages.model.FPGA.FPGA_N_NE;
-import static uk.ac.manchester.spinnaker.messages.model.FPGA.FPGA_SW_W;
 import static uk.ac.manchester.spinnaker.messages.model.FPGAMainRegisters.LEDO;
 import static uk.ac.manchester.spinnaker.messages.model.FPGAMainRegisters.SCRM;
 import static uk.ac.manchester.spinnaker.messages.model.FPGAMainRegisters.SLEN;
@@ -62,7 +62,7 @@ import org.springframework.stereotype.Component;
 import uk.ac.manchester.spinnaker.alloc.model.Prototype;
 import uk.ac.manchester.spinnaker.machine.MemoryLocation;
 import uk.ac.manchester.spinnaker.machine.board.BMPBoard;
-import uk.ac.manchester.spinnaker.messages.model.FPGA;
+import uk.ac.manchester.spinnaker.machine.board.FPGA;
 import uk.ac.manchester.spinnaker.messages.model.FPGALinkRegisters;
 import uk.ac.manchester.spinnaker.messages.model.FPGAMainRegisters;
 import uk.ac.manchester.spinnaker.transceiver.BMPTransceiverInterface;
