+++ SpiNNaker-comms/src/test/java/uk/ac/manchester/spinnaker/messages/scp/GeneralMessageTest.java
@@ -22,8 +22,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static uk.ac.manchester.spinnaker.machine.ChipLocation.ZERO_ZERO;
 import static uk.ac.manchester.spinnaker.machine.Direction.EAST;
 import static uk.ac.manchester.spinnaker.machine.MemoryLocation.NULL;
+import static uk.ac.manchester.spinnaker.machine.board.FPGA.FPGA_E_S;
 import static uk.ac.manchester.spinnaker.messages.model.CPUState.IDLE;
-import static uk.ac.manchester.spinnaker.messages.model.FPGA.FPGA_E_S;
 import static uk.ac.manchester.spinnaker.messages.model.IPTagTimeOutWaitTime.TIMEOUT_10_ms;
 import static uk.ac.manchester.spinnaker.messages.model.LEDAction.TOGGLE;
 import static uk.ac.manchester.spinnaker.messages.model.PowerCommand.POWER_OFF;
