+++ SpiNNaker-machine/src/main/java/uk/ac/manchester/spinnaker/machine/board/FPGA.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package uk.ac.manchester.spinnaker.messages.model;
+package uk.ac.manchester.spinnaker.machine.board;
 
 /**
  * Identifiers for the FPGAs on a SpiNNaker board, as managed by BMP.
