+++ SpiNNaker-machine/src/main/java/uk/ac/manchester/spinnaker/machine/board/doc-files/Directions.fig
